+++ processor/src/test/fixtures/input/com/example/computable_value_ref/ProtectedAccessFromBaseComputableValueRefModel.java
@@ -1,0 +1,10 @@
+package com.example.computable_value_ref;
+
+import arez.annotations.ArezComponent;
+import com.example.computable_value_ref.other.BaseProtectedAccessComputableValueRefModel;
+
+@ArezComponent
+public abstract class ProtectedAccessFromBaseComputableValueRefModel
+  extends BaseProtectedAccessComputableValueRefModel
+{
+}