+++ processor/src/test/fixtures/input/com/example/computable_value_ref/Suppressed2ProtectedAccessComputableValueRefModel.java
@@ -1,0 +1,24 @@
+package com.example.computable_value_ref;
+
+import arez.ComputableValue;
+import arez.annotations.ArezComponent;
+import arez.annotations.ComputableValueRef;
+import arez.annotations.Memoize;
+import arez.annotations.SuppressArezWarnings;
+import javax.annotation.Nonnull;
+
+@ArezComponent
+public abstract class Suppressed2ProtectedAccessComputableValueRefModel
+{
+  @Memoize
+  public long getTime()
+  {
+    return 0;
+  }
+
+  // This uses the CLASS retention suppression
+  @SuppressArezWarnings( "Arez:ProtectedRefMethod" )
+  @Nonnull
+  @ComputableValueRef
+  protected abstract ComputableValue<Long> getTimeComputableValue();
+}