+++ processor/src/test/fixtures/input/com/example/computable_value_ref/other/BaseProtectedAccessComputableValueRefModel.java
@@ -1,0 +1,20 @@
+package com.example.computable_value_ref.other;
+
+import arez.ComputableValue;
+import arez.annotations.ArezComponent;
+import arez.annotations.ComputableValueRef;
+import arez.annotations.Memoize;
+import javax.annotation.Nonnull;
+
+public abstract class BaseProtectedAccessComputableValueRefModel
+{
+  @Memoize
+  public long getTime()
+  {
+    return 0;
+  }
+
+  @Nonnull
+  @ComputableValueRef
+  protected abstract ComputableValue<Long> getTimeComputableValue();
+}